Add and install Clerk for user management
After a lot of trial and error with Next Auth, I've decided that it's
best for my sanity and for the survival of this project to drop it until
they're in a better place with regards to modern Next (ie the App
Router). This will be used just for authentication initially, but I'll
be able to set up webhooks to create records on sign in. It's not as
hands on as Next Auth, but the experience of setting it up has been
awesome so far and I'm excited to use this more!