Add Task page for editing task values
This change adds a page to edit the attributes of a task, as well as any
recurrence values it might have. Not the cleanest React code I've
written, but it gets the job done and gets me back to working on much
simpler pieces of functionality. I don't have plans to create anything
more complicated than the recurrence form, so hopefully with more
experience I can come back to this in a year and refactor into it into
something nicer.