Rename Task Completions to Task Assignments
It's just an obviously better name, I wish I'd taken more time to think
of it but I must've been really burned out when I was writing this code.
Huge diff because a lot of references were changed along with file names
and directories, but no major functionality. There is a corresponding PR
in the backend codebase to change the Rails model and routes.